Update block btree
There's a lot of evolution of the block btree here.

We move to a fixed size key that has 3 u64s, mostly to simplify key
access throughout.  We add a small number of key functions for setting,
comparing, incrementing, etc.

We extend the item offset array to be a header that also includes the
item's value size.  This makes it easy to align the keys and values
without losing space to padding.

We switch to read and write iterator callbacks for read and write access
to items, and we fully implement tree iteration, growth, and shrinking.

And though all this we use the transaction store wrappers to work with
the undo buffers.

Signed-off-by: Zach Brown <zab@versity.com>