Rework userspace mount and debugfs user
The setup of the debugfs thread was (and probably still is!) pretty
clumsy.  We have option parsing in the mount helper.  That includes the
trace file name so we setup tracing after option parsing.  But the
debugfs cli command called thread_prepare_main, which registers itself
as a tracing thread and wants tracing setup, before running the debugfs
parser thread which was mounting and setting up tracing.
mess.

The attempt here is to move thread_prepare_main into the mount helper.
This imposes a main/thread pattern on processes that call mount to
setup.  With this in place the debugfs cli command mounts before
launching the debugfs command parser thread.

We'll see if this holds.

(While we're at it we drop the silly manual test for required devd
address options and mark the option as required.)

Signed-off-by: Zach Brown <zab@versity.com>