Fix atomic_inc_unless_negative()
atomic_inc_unless_negative() is intended to implement the acquistion
of a read lock on a shared reader/exclusive writer lock. The value of
the atomic counter is either positive, meaning that number of readers
are active, or negative, meaning that there is a single writer. A read
lock is acquired by reading the value of the lock, and if it is
positive, adding one to the value and doing a cmpxchg. The cmpxchg
returns the value of the counter at the time the cmpxchg was
attempted. If we acquired the reader lock, then this value should be
the same as the value we originally read. Due to a copy-and-paste
error, it was instead checking if the value was 0 - which works until
there is a long-lived read reference, such as while a block is being
flushed. Effectively it was an exclusive reader lock. Fix the
comparison to be against the previous value, not 0.

Signed-off-by: Valerie Aurora <val@versity.com>
Signed-off-by: Zach Brown <zab@zabbo.net>