Fix free/alloc reversal of btree block merge condition
To achieve the goal of merging two btree blocks when the merged block
would be 80% full, we want to merge two blocks when their _used_ space
is 40% or lower, not when their _free_ space is 40% or lower. Add some
BUG_ON()'s to catch total_free wrapping around to negative, which was
the visible symptom in this case.

Signed-off-by: Valerie Aurora <val@versity.com>
Signed-off-by: Zach Brown <zab@zabbo.net>