Remove block transport
The block cache implementation shared by the kernel client and userspace
servers was built with pluggable block IO transports.  Clients would
have the networking transport and devd would have a local block
transport.

Now that the userspace server side is using a very different runtime
they won't be reusing the block cache from the client.  We can simplify
it quite a bit by having it use messaging directly.

This is in preparation for adding the cache mode negotation with devd
which wasn't worth trying to squeeze through the block transport
interface.

Signed-off-by: Zach Brown <zab@zabbo.net>