Add cache mode to shared client block cache
This adds awareness of cache modes to the multi-threaded client block
cache.

The big change in the code is to move from operating on individual state
bits to atomic changes of a multi-state struct.  This makes it trivial
to atomically make complicated state transitions without intermediate
combinations of states being exposed to readers.  Arguably, it also
makes the code a lot nicer to follow.

With that in place, we can add cache mode changes along with the other
state changes to cached blocks.

As a read request is being sent to initially populate the cache, we
include a requested cache mode that will satisfy the _get caller that
triggered the read.  The read result includes the cache mode that is
granted.  From there, further _get calls can use the same pattern to
request further modes that were incompatible with the first (requesting
write after populating to read).

The more involved changes are respondig to unsolicited messages which
change the cache mode.  We have to wait for incompatible users to drain
before we can send the ack.  And if the block was dirty, we may have to
also write out its contents.

Signed-off-by: Zach Brown <zab@zabbo.net>