Send quit signal to process, not process group
The quit command in debugfs was sending a signal to every process in
the process group (using 0 for pid argument of kill()), which breaks
the ngnfs tests scripts because the shell script running debugfs is
part of the same process group and gets the signal too. Send the
signal only to this process, instead.

Signed-off-by: Valerie Aurora <val@versity.com>
Signed-off-by: Zach Brown <zab@zabbo.net>