Import kernel's generic crc64 to our lk
Add support for 64bit crc via the kernel's routines.  The generic
implemention is built around a binary that generates tables.  This keeps
support for both the ECMA and NVME polynomials, though we'll only be
using the NVME variant.  This doesn't bother importing the x86-specific
SIMD implementation, though we'll probably want to use it eventually.

Signed-off-by: Zach Brown <zab@zabbo.net>