Significantly refactor utask block cache
We're about to add transactional writes to devd.  It will be responsible
for managing concurrent read or write access to blocks as they're
written to the journal, or back out to the main block store, or what
have you.

This puts less pressure on the behavioural requirements of the utask
block cache, of which devd is the only user.  The big change is that the
block cache doesn't need to worry about read/write exclusion.  We remove
the funky readers and writer counters and replace them with a familiar
refcount.  The freeing path is changed from a fiddly state check that
gates presence on the lru to simply executing when the refcount drops to
zero.  Blocks hold elevated refcounts when they're on lists or under IO,
and of course callers hold references.

So with the refcount change, it becomes a lot cleaner to implement a
simple fifo for reclaim, a preallocated pool of blocks, a list of dirty
blocks, and we track blocks in a hash table instead of the rbtree
because we don't need sorting.

Finally we add some trivial conveniences to return the total count of
blocks covered by the cache and to clear a block pointer after being
put.

Signed-off-by: Zach Brown <zab@zabbo.net>