Support utask cancel in utask net
The utask net subsystem didn't have support for shutting down and
destroying all existing sockets.  It would only destroy sockets if any
of its service utasks saw an error, and then utask_destroy would be
called from that utask.

the current model is to call utask_destroy from roughly the same context
that calls _create.  Usually that's subsystem init and exit.  But the
net layer has utasks created to service specific sockets.

Instead of destroying sockets from within their service utasks, this
adds a global utask that is responsible for destroying sockets.  Sockets
are added to a list that it consumes.  We add a global list of sockets
that are spliced onto the destruction list as we shut down the
subsystem.

There's an often repeated pattern where we bubble up canceled errors
from the waiter utask blocking wrappers around io_uring networking
requests.

This leaves the tree a bit broken because devd doesn't call net_init or
net_exit.  That's up next.

Signed-off-by: Zach Brown <zab@zabbo.net>