Rework devd main to run in utask
The devd blocking store reads the journal as it initializes.  The
implementation re-uses its normal block IO functions that use the utask
block cache.  This necessitates running the bstore subsystem init from a
utask.

We take the opportunity to shift most of the devd initialization into a
utask.  This makes it less special than the rest of the code.  We can
use utask and io_uring constructs to implement orderly shutdown.  The
only steps in the main process context are setting up and launching the
utask.

Signed-off-by: Zach Brown <zab@zabbo.net>