Catch errors in calls to ngnfs_tblk_zero_tail()
ngnfs_tblk_zero_tail() was testing for the head (area to not zero)
being larger than than the tail by subtracting and checking for a
negative result and silently doing nothing. But the variables are the
same type as those being passed in, which are always unsigned, so the
test never triggered. Instead, assert that the head is less than or
equal to the tail. head == tail should be allowed since the main use
is zeroing out the difference between the unaligned length and aligned
length of a btree item, which are sometimes the same.

Signed-off-by: Valerie Aurora <val@versity.com>
Signed-off-by: Zach Brown <zab@zabbo.net>