Have devd print tasks on SIGUSR1
Add a quick hack to start getting some visibility into utasks.

We add an internal utask tracking list so that the core can iterate over
all the allocated tasks.  Previusly utask_schedule() was the asm
fragment that performed the stack switcheroo.  We rename it to give it
symmetry with _switch_to().  That lets us offer utask_schedule() as a
macro and function that can update task info before scheduling away.
Here we record the call site of the schedule.  We give task creation a
similar treatment so that we can name each task by its function name.

Finally we offer a call to print all the utasks to stderr.  We wire it
up to receiving SIGUSR1 in devd.

Signed-off-by: Zach Brown <zab@zabbo.net>