Add support for valgrind
Add some support for valgrind to help it understand what we're doing.

Registering our stacks stops it from warning when it sees the stack
pointers change by more than its max tolerance.

It also doesn't understand that io_uring commands are writing to memory
regions.  We add a helper that initializes buffers that we're writing to
with io_uring commands.  This stops *A LOT* of uninialized use warnings.
A lot of the block read commands don't need initialization because
they're already zeroing the buffers.  This could be cleaned up a bit.

We always build in support at the moment which adds a build dependency
on the valgrind headers.

Signed-off-by: Zach Brown <zab@zabbo.net>