Add block details to devd bstore read and write
The internals of the bstore_{read,write} methods had support for working
with the block detail struct.  _read() had it available and _write() was
updating it.  But it wasn't exposed to the callers.

This adds the detail struct to the read and write calls so that the
reader can get access to the details and the writer can update them.

Signed-off-by: Zach Brown <zab@zabbo.net>