Update devd block cache message handling
The kernel client is coming online and it requires a bit more rigorous
block cache protocol implementation.  This improves the devd cache mode
handling server path in a few directions.

The amount of memory allocated to track each cached block is on the way
to being a lot more efficient.  Per-client memory use went from a big
rbtree node down to a small array entry.  (It'll be much smaller once
the system evolves to have maps with assigned client IDs.).

The protocol allows for back-to-back sending of a few key state
transitions.  The client can send requests for read and then write mode
without waiting for the response to the read.  Similarly, the server can
send mode settting invalidation messages for read, then none.  This
reduces latency and extra messages on the wire.  We avoid some
complexity by requiring in-order processing through the messaging layer.

The mechanism for using a block read message as a mode request message
is made a bit clearer.  We put the read in the processing loop to make
it more clear what's going on, rather than hiding it in the sending
message.  This has the benefit of letting a processing loop send the
same block contents to multiple waiting read requests for the same
block.

Finally, a bit of block metadata (the write count) is plumbed between
the read and write messages and the bstore IO API.  This supports the
block transaction engine in the client needing to know when cached
blocks changed while it was working.

Signed-off-by: Zach Brown <zab@zabbo.net>