Restore header defines for older sparse
As gcc's standard support has increased, it has taken on native support
that used to be provided by defines.  When sparse doesn't support the
functionality without the defines we hit errors.

An example is native support for bool/true/false without being defined
to _Bool/1/0.  Ass gcc gained support for those keywords it dropped the
defines.  Now the code that sparse is checking has unknown keywords
instead of the result of the definitions.

There are a handful of these cases that we've hit errors in.  This looks
at the C standard that gcc is building against, and when the headers
emit code that sparse doesn't understand we restore the defines that the
headers used to.

Signed-off-by: Zach Brown <zab@zabbo.net>