Make min device size more visible
The minimum device size was arbitrarily enforced by devd's bstore as it
started up.  Record that (still arbitrary) minimum device size, in terms
of the number of commit or journal blocks in the format header.  Check
that the device is large enough when formatting, and add tracing so we
can at least see the layout when bstore starts up and exits.

Signed-off-by: Zach Brown <zab@zabbo.net>