Track client cache modes in an rbtree
Previously client cache modes were tracked in a per-block structure with
an array recording each client's state for that block.  This switches
the data structure to have structs in an rbtree for each cached block at
each client.

This is in preparation for supporting bulk grants for allocation.  We
want to be able to grant write mode to many blocks at a client.  This
simpler structure is a lot easier to work with.

A lot of the high level logic remains the same.  The iterator that
previously walked the client array in a block struct now walks client
block structs.

The process for determining the next pending request does change.  It
used to use a rotating index to decide the next client request to be
processed for a block.  We no longer have an index into rbtree nodes, so
we visit all the client blocks and randomly chose the next pending
request to process.

Signed-off-by: Zach Brown <zab@zabbo.net>